Organização da parte Visual do projeto
A classe contendo as telas do programa foi organizada em seu pacote e codificada em utf-8
